# 实训打卡.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "实训打卡",
    "description": "",
    "version": "1.0.0"
  },
...
  "paths": {
...
    "/users/me/checkin-records": {
      "get": {
        "summary": "获取当前用户签到记录",
        "deprecated": false,
        "description": "获取当前用户的所有签到记录",
        "tags": [
          "CheckinRecords"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "成功返回用户的签到记录列表",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessWithData"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CheckinRecord"
                          }
                        }
                      }
                    }
                  ]
                },
                "examples": {
                  "签到记录": {
                    "summary": "用户签到记录列表示例",
                    "value": {
                      "code": 0,
                      "data": [
                        {
                          "recordId": 98765,
                          "taskId": 2023081501,
                          "taskName": "早会签到",
                          "groupId": 50,
                          "groupName": "技术一组",
                          "userId": 101,
                          "username": "zhangsan",
                          "signedTime": 1689486600,
                          "locationInfo": {
                            "location": {
                              "latitude": 39.915,
                              "longitude": 116.404
                            }
                          },
                          "status": "success",
                          "message": "签到成功",
                          "checkinMethods": {
                            "gps": true,
                            "face": true,
                            "wifi": false,
                            "nfc": false
                          },
                          "createdAt": 1689486600
                        },
                        {
                          "recordId": 98766,
                          "taskId": 2023081405,
                          "taskName": "项目进度汇报",
                          "groupId": 51,
                          "groupName": "项目组A",
                          "userId": 101,
                          "username": "zhangsan",
                          "signedTime": 1689399900,
                          "locationInfo": {
                            "location": {
                              "latitude": 39.917,
                              "longitude": 116.406
                            }
                          },
                          "status": "failed_location",
                          "message": "位置不符",
                          "failureReason": "不在签到范围内",
                          "checkinMethods": {
                            "gps": true,
                            "face": false,
                            "wifi": true,
                            "nfc": false
                          },
                          "createdAt": 1689399900
                        }
                      ]
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "用户未登录或提供的Token无效/过期",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器内部错误，获取签到记录时发生异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/checkin-tasks/{taskId}/checkin": {
      "post": {
        "summary": "执行签到",
        "deprecated": false,
        "description": "用户针对某个签到任务执行签到操作。根据任务配置的校验方式提供相应数据。需要登录且是任务所属组成员。",
        "tags": [
          "CheckinRecords"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "签到任务 ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int",
              "x-oapi-codegen-extra-tags": {
                "uri": "taskId",
                "binding": "required,gt=0"
              }
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "verificationData": {
                    "$ref": "#/components/schemas/VerificationData",
                    "description": "签到校验数据，根据verifyType提供对应字段的数据"
                  },
                  "signinTime": {
                    "type": "integer",
                    "format": "int",
                    "description": "用户提交的签到时间（Unix时间戳，单位：秒），默认为当前时间",
                    "x-go-type-skip-optional-pointer": true,
                    "x-oapi-codegen-extra-tags": {
                      "json": "signinTime",
                      "binding": "omitempty"
                    }
                  },
                  "groupId": {
                    "type": "integer",
                    "format": "int",
                    "description": "用户组ID",
                    "x-go-type-skip-optional-pointer": true,
                    "x-oapi-codegen-extra-tags": {
                      "json": "groupId",
                      "binding": "omitempty,gt=0"
                    }
                  }
                },
                "required": [
                  "verificationData"
                ],
                "description": "请根据任务的校验方式配置提供相应的校验数据。服务器会根据任务配置校验必要字段。"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "签到请求处理完成，返回签到结果状态（包括成功、失败及原因、签到时间状态）",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessWithData"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "object",
                          "properties": {
                            "recordId": {
                              "type": "integer",
                              "format": "int",
                              "description": "签到记录ID"
                            },
                            "signedTime": {
                              "type": "integer",
                              "format": "int",
                              "description": "签到时间戳（秒）"
                            },
                            "success": {
                              "type": "boolean",
                              "description": "签到是否成功"
                            }
                          },
                          "required": [
                            "recordId",
                            "signedTime",
                            "success"
                          ]
                        }
                      }
                    }
                  ]
                },
                "example": {
                  "code": 0,
                  "data": {
                    "recordId": 98765,
                    "signedTime": 1689486600,
                    "success": true
                  }
                }
              }
            },
            "headers": {}
          },
          "400": {
            "description": "请求参数错误，可能原因：缺少必需参数、参数格式不正确或缺少任务配置所需的校验数据",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequest"
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "认证失败，需要重新登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            },
            "headers": {}
          },
          "403": {
            "description": "权限不足，可能原因：不是任务所属组的成员",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "请求的任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器内部错误，处理签到请求时发生异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
    "/checkin-tasks/{taskId}/records": {
      "get": {
        "summary": "获取签到任务的签到记录列表 (管理员视角)",
        "deprecated": false,
        "description": "用户组管理员查看某个签到任务的所有成功签到记录。",
        "tags": [
          "CheckinRecords"
        ],
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "description": "签到任务 ID",
            "required": true,
            "example": 0,
            "schema": {
              "type": "integer",
              "format": "int",
              "x-oapi-codegen-extra-tags": {
                "uri": "taskId",
                "binding": "required,gt=0"
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取指定任务的所有成功签到记录",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/SuccessWithData"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/CheckinRecord"
                          },
                          "description": "所有成功签到的成员记录"
                        }
                      }
                    }
                  ]
                },
                "examples": {
                  "签到记录列表": {
                    "summary": "任务签到记录示例",
                    "value": {
                      "code": 0,
                      "data": [
                        {
                          "recordId": 98765,
                          "taskId": 2023081501,
                          "taskName": "每日站会签到",
                          "groupId": 50,
                          "groupName": "研发一部",
                          "userId": 101,
                          "username": "张三",
                          "signedTime": 1689486600,
                          "message": "签到成功",
                          "checkinMethods": {
                            "gps": true,
                            "face": true,
                            "wifi": false,
                            "nfc": false
                          }
                        },
                        {
                          "recordId": 98770,
                          "taskId": 2023081501,
                          "taskName": "每日站会签到",
                          "groupId": 50,
                          "groupName": "研发一部",
                          "userId": 105,
                          "username": "钱七",
                          "signedTime": 1689486800,
                          "message": "签到成功",
                          "checkinMethods": {
                            "gps": true,
                            "face": true,
                            "wifi": false,
                            "nfc": false
                          }
                        }
                      ]
                    }
                  }
                }
              }
            },
            "headers": {}
          },
          "401": {
            "description": "认证失败，需要重新登录",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Unauthorized"
                }
              }
            },
            "headers": {}
          },
          "403": {
            "description": "权限不足，可能原因：不是任务所属组的管理员",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Forbidden"
                }
              }
            },
            "headers": {}
          },
          "404": {
            "description": "请求的任务不存在",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFound"
                }
              }
            },
            "headers": {}
          },
          "500": {
            "description": "服务器内部错误，获取签到记录列表时发生异常",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InternalServerError"
                }
              }
            },
            "headers": {}
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CheckinRecord": {
        "type": "object",
        "properties": {
          "recordId": {
            "type": "integer",
            "format": "int",
            "description": "签到记录ID",
            "readOnly": true,
            "x-go-type-skip-optional-pointer": true,
            "examples": [
              98765
            ]
          },
          "taskId": {
            "type": "integer",
            "format": "int",
            "description": "签到任务ID",
            "readOnly": true,
            "x-go-type-skip-optional-pointer": true,
            "examples": [
              2023081501
            ]
          },
          "taskName": {
            "type": "string",
            "description": "签到任务名称",
            "readOnly": true,
            "x-go-type-skip-optional-pointer": true,
            "examples": [
              "每日站会签到"
            ]
          },
          "groupId": {
            "type": "integer",
            "format": "int",
            "description": "用户组ID",
            "readOnly": true,
            "x-go-type-skip-optional-pointer": true,
            "examples": [
              50
            ]
          },
          "groupName": {
            "type": "string",
            "description": "用户组名称",
            "readOnly": true,
            "x-go-type-skip-optional-pointer": true,
            "examples": [
              "研发一部"
            ]
          },
          "userId": {
            "type": "integer",
            "format": "int",
            "description": "签到用户ID",
            "readOnly": true,
            "x-go-type-skip-optional-pointer": true,
            "examples": [
              102
            ]
          },
          "username": {
            "type": "string",
            "description": "签到用户名",
            "readOnly": true,
            "x-go-type-skip-optional-pointer": true,
            "examples": [
              "lisi"
            ]
          },
          "signedTime": {
            "type": "integer",
            "format": "int",
            "description": "实际签到时间（Unix时间戳，单位：秒）",
            "readOnly": true,
            "x-go-type-skip-optional-pointer": true,
            "examples": [
              1689486600
            ]
          },
          "message": {
            "type": "string",
            "description": "签到信息",
            "x-go-type-skip-optional-pointer": true,
            "examples": [
              "签到成功"
            ]
          },
          "checkinMethods": {
            "$ref": "#/components/schemas/CheckinMethods",
            "description": "该次签到使用的校验方式组合",
            "readOnly": true
          },
          "locationInfo": {
            "type": "object",
            "properties": {
              "location": {
                "$ref": "#/components/schemas/Location"
              }
            },
            "description": "位置信息"
          },
          "createdAt": {
            "type": "integer",
            "format": "int",
            "description": "创建时间（Unix时间戳，单位：秒）",
            "readOnly": true,
            "x-go-type-skip-optional-pointer": true,
            "examples": [
              1689486600
            ]
          }
        },
        "required": [
          "recordId",
          "taskId",
          "taskName",
          "groupId",
          "groupName",
          "userId",
          "username",
          "checkinMethods"
        ],
        "description": "签到记录，包含任务基本信息和签到信息。只包含成功签到的记录。"
      },
...
      "VerificationData": {
        "type": "object",
        "properties": {
          "locationInfo": {
            "type": "object",
            "properties": {
              "location": {
                "$ref": "#/components/schemas/Location"
              }
            },
            "required": [
              "location"
            ],
            "description": "位置校验信息（仅当任务需要GPS校验时必须提供）"
          },
          "faceData": {
            "type": "string",
            "format": "byte",
            "description": "人脸识别数据的Base64编码（仅当任务需要人脸校验时必须提供）",
            "x-go-type-skip-optional-pointer": true
          },
          "wifiInfo": {
            "$ref": "#/components/schemas/WifiInfo",
            "description": "检测到的WiFi信息（仅当任务需要WiFi校验时必须提供）"
          },
          "nfcInfo": {
            "$ref": "#/components/schemas/NFCInfo",
            "description": "扫描到的NFC标签信息（仅当任务需要NFC校验时必须提供）"
          }
        },
        "description": "校验数据组件，根据不同的校验方式需要提供不同的字段"
      },
...
      "SuccessWithData": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "data": {
                "oneOf": [
                  {
                    "type": "object",
                    "description": "具体数据结构需在各API中定义",
                    "properties": {}
                  },
                  {
                    "type": "array",
                    "description": "具体数据结构需在各API中定义",
                    "items": {
                      "type": "string"
                    }
                  }
                ]
              }
            },
            "required": [
              "data"
            ]
          }
        ]
      },
...
      "Unauthorized": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "message"
            ]
          }
        ]
      },
      "Forbidden": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "message"
            ]
          }
        ]
      },
      "NotFound": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "message"
            ]
          }
        ]
      },
...
      "BadRequest": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "message"
            ]
          }
        ]
      },
      "InternalServerError": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseResponse"
          },
          {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "message"
            ]
          }
        ]
      }
    },
    "securitySchemes": {
      "JWT鉴权": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "servers": [],
  "security": []
}